+++ src/main/java/cart/application/ProductService.java
@@ -1,7 +1,7 @@
 package cart.application;
 
 import cart.dao.ProductDao;
-import cart.domain.Product;
+import cart.domain.product.Product;
 import cart.dto.request.ProductRequest;
 import cart.dto.response.ProductResponse;
 import org.springframework.stereotype.Service;
