+++ src/main/java/cart/domain/member/Email.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.member;
 
 import java.util.Objects;
