+++ src/main/java/cart/domain/member/Password.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.member;
 
 import org.springframework.util.StringUtils;
