+++ src/main/java/cart/dto/response/MemberResponse.java
@@ -1,6 +1,6 @@
 package cart.dto.response;
 
-import cart.domain.Member;
+import cart.domain.member.Member;
 
 public class MemberResponse {
     private final Long id;
