+++ src/main/java/cart/ui/MemberArgumentResolver.java
@@ -1,7 +1,7 @@
 package cart.ui;
 
 import cart.dao.MemberDao;
-import cart.domain.Member;
+import cart.domain.member.Member;
 import cart.exception.AuthenticationException;
 import org.apache.tomcat.util.codec.binary.Base64;
 import org.springframework.core.MethodParameter;
