+++ src/test/java/cart/domain/CartItemTest.java
@@ -1,5 +1,8 @@
 package cart.domain;
 
+import cart.domain.cart.CartItem;
+import cart.domain.member.Member;
+import cart.domain.product.Product;
 import cart.exception.CartItemException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
