+++ src/test/java/cart/domain/MemberTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.member.Member;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
