+++ src/test/java/cart/domain/OrderProductsTest.java
@@ -1,5 +1,9 @@
 package cart.domain;
 
+import cart.domain.cart.CartItem;
+import cart.domain.member.Member;
+import cart.domain.order.OrderProducts;
+import cart.domain.product.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
