+++ src/test/java/cart/integration/CartItemIntegrationTest.java
@@ -2,8 +2,8 @@ package cart.integration;
 
 import cart.dao.MemberDao;
 import cart.dao.ProductDao;
-import cart.domain.Member;
-import cart.domain.Product;
+import cart.domain.member.Member;
+import cart.domain.product.Product;
 import cart.dto.request.CartItemQuantityUpdateRequest;
 import cart.dto.request.CartItemRequest;
 import cart.dto.request.PaymentRequest;
