+++ src/test/java/cart/integration/MemberIntegrationTest.java
@@ -1,7 +1,7 @@
 package cart.integration;
 
 import cart.dao.MemberDao;
-import cart.domain.Member;
+import cart.domain.member.Member;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
